backend/requirements.txt
flask
flask-cors
requests
transformers
torch
gunicorn
python-dotenv
langchain
langchain-groq
flask-limiter